+++ src/Setups/SetupConfigs/Players/AllPlayers.ts
@@ -1,4 +1,4 @@
-import * as DefaultPlayers_0 from "./DefaultPlayers_0";
+import * as DefaultPlayers_0 from "./RaidPlayers";
 import * as DefaultPlayers_1 from "./DefaultPlayers_1";
 import * as DefaultPlayers_4 from "./DefaultPlayers_4";
