+++ src/Setups/SetupConfigs/Setups/QadimCC.tsx
@@ -1,7 +1,7 @@
 import {RaidSetup} from "./RaidSetup";
-import {Player} from "./DefaultPlayers_1";
+import {Player} from "../Players/DefaultPlayers_1";
 
-const DhuumCM : RaidSetup<Player> = {
+const QadimCC : RaidSetup<Player> = {
             "playerSettings": {
                 "players": ["Kalell", "Nevra", "Yasi", "Alex", "Lumi", "Fynn", "David", "Shrotty", "Assa", "Ely"],
                 "missing": [],
@@ -9,19 +9,24 @@ const DhuumCM : RaidSetup<Player> = {
             },
             "list": [
                 {
-                    "name": "Wing 5",
+                    "name": "Wing 6",
                     "bosses": [
                         {
-                            "name": "Dhuum CM",
-                            "iconMiniId": 85998,
+                            "name": "Qadim",
+                            "iconMiniId": 88587,
                             "selectedSetup": 0,
                             "setups": [
                                 {
                                     "name": "Normal",
                                     "roles": [
                                         {
+                                            "profession": "Chronomancer",
+                                            "tasks": ["Tank", "Quickness", "Portal"],
+                                            "player": "Kalell"
+                                        },
+                                        {
                                             "profession": "Tempest",
-                                            "tasks": ["Tank", "Heal", "Buffs"],
+                                            "tasks": ["Heal", "Buffs"],
                                             "player": "Shrotty"
                                         },
                                         {
@@ -30,29 +35,19 @@ const DhuumCM : RaidSetup<Player> = {
                                             "player": "Ely"
                                         },
                                         {
-                                            "profession": "Firebrand",
-                                            "tasks": ["Full Quickness"],
+                                            "profession": "Daredevil",
+                                            "tasks": ["Portal"],
                                             "player": "Yasi"
                                         },
                                         {
-                                            "profession": "Berserker",
-                                            "tasks": ["Green 1"],
-                                            "player": "David"
+                                            "profession": "Daredevil",
+                                            "tasks": ["Solo Kiter"],
+                                            "player": "Alex"
                                         },
                                         {
-                                            "profession": "Scourge",
-                                            "tasks": ["Kiter", "Green 2"],
-                                            "player": "Kalell"
-                                        },
-                                        {
-                                            "profession": "Chronomancer",
-                                            "tasks": ["Green 3"],
+                                            "profession": "Reaper",
+                                            "tasks": ["Lampe"],
                                             "player": "Nevra"
-                                        },
-                                        {
-                                            "profession": "Firebrand",
-                                            "tasks": ["First Green 2"],
-                                            "player": "Lumi"
                                         },
                                     ]
                                 },
@@ -63,4 +58,4 @@ const DhuumCM : RaidSetup<Player> = {
             ]
         };
 
-export default DhuumCM;
+export default QadimCC;
