+++ src/raid/Boss.tsx
@@ -3,7 +3,7 @@ import ItemIcon from "../helper/ItemIcon";
 
 import './Boss.scss';
 import SetupSelector from "./SetupSelector";
-import * as RaidSetup from "../Setups/SetupConfigs/RaidSetup";
+import * as RaidSetup from "../Setups/SetupConfigs/Setups/RaidSetup";
 
 interface ISetup {
     props : {setupValue: {name: string}}
